Debug mf iga for 64 elements

+++ src/iga_wq_mf/ThermoMechaIGA/pysrc/lib/python_iga.py
@@ -18,7 +18,7 @@ class IGA(thermoMechaModel):
         super().__init__(modelIGA, material=material, Dirichlet=Dirichlet)
 
         # Evaluate basis and weights
-        self._nb_qp, self._nb_qp_total = np.ones(3), None
+        self._nb_qp, self._nb_qp_total = np.ones(3, dtype=int), None
         self.eval_basis_weights()
 
         # Get jacobian and physical position
